Proper Open and Save dialogs. #167
